--- src/main/java/lotto/NumberMatcher.java
@@ -1,7 +1,0 @@
-package lotto;
-
-public interface NumberMatcher {
-
-    boolean match(Integer number);
-
-}
