+++ src/main/java/lotto/domain/Lotto.java
@@ -1,7 +1,5 @@
 package lotto.domain;
 
-import java.io.PrintWriter;
-
 public class Lotto {
 
     public static final int PRICE = 1000;
@@ -12,11 +10,7 @@ public class Lotto {
         this.number = number;
     }
 
-    public long getMatchedNumberCount(NumberMatcher numberMatcher) {
-        return this.number.getMatchedNumberCount(numberMatcher);
-    }
-
-    public void printNumber(PrintWriter writer) {
-        this.number.print(writer);
+    public int getMatchedNumberCount(NumberMatcher numberMatcher) {
+        return number.getMatchedNumberCount(numberMatcher);
     }
 }
