+++ src/main/java/lotto/domain/LottoNumberGenerator.java
@@ -1,31 +1,25 @@
 package lotto.domain;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Random;
+import java.util.*;
 import java.util.stream.IntStream;
 
 public interface LottoNumberGenerator {
 
     /**
      * LottoNumber 를 생성합니다
-     *
      * @return
      */
     LottoNumber generate();
 
-    static LottoNumberGenerator simple() {
+    static LottoNumberGenerator simple(){
 
         return new LottoNumberGenerator() {
 
             private ArrayList<Integer> lottoNumberBalls;
-
             {
                 lottoNumberBalls = new ArrayList<>();
-                IntStream.rangeClosed(1, 45).forEach(lottoNumberBalls::add);
+                IntStream.rangeClosed(1, 45).forEach( lottoNumberBalls::add );
             }
-
             private Random random = new Random();
 
             @Override
