+++ src/test/java/lotto/domain/LottoMachineTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.domain.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -18,7 +19,7 @@ public class LottoMachineTest {
 
     @BeforeEach
     void setUp() {
-        this.prizeMoneys = Arrays.asList(2000000000, 1500000, 50000, 2000);
+        this.prizeMoneys = Arrays.asList(10000, 5000, 3000, 1000);
         this.lottoMachine = new LottoMachine(this.prizeMoneys);
     }
 
@@ -33,11 +34,11 @@ public class LottoMachineTest {
     @Test
     void rateOfReturn() {
         Lotto secondPrizeLotto = myLotto(1, 2, 3, 4, 5, 10);
-        WinningNumber winningNumber = new WinningNumber("1, 2, 3, 4, 5, 6");
+        WinningNumber winningNumber = new WinningNumber(Arrays.asList(1, 2, 3, 4, 5, 6));
 
-        PrizeWinningResult result = lottoMachine.checkPrizeWinning(winningNumber, secondPrizeLotto);
+        PrizeWinningResult result = lottoMachine.checkPrizeWinning(new Lottos(secondPrizeLotto), winningNumber);
 
-        assertThat(result.getRateOfReturn()).isEqualTo(1500);
+        assertThat(result.getRateOfReturn()).isEqualTo(5);
     }
 
 
@@ -47,20 +48,20 @@ public class LottoMachineTest {
         Lotto firstPrizeLotto = myLotto(1, 2, 3, 4, 5, 6);
         Lotto fouthPrizeLotto = myLotto(1, 2, 3, 10, 20, 30);
 
-        WinningNumber winningNumber = new WinningNumber("1, 2, 3, 4, 5, 6");
+        WinningNumber winningNumber = new WinningNumber(Arrays.asList(1, 2, 3, 4, 5, 6));
 
-        PrizeWinningResult result = lottoMachine.checkPrizeWinning(winningNumber, firstPrizeLotto, fouthPrizeLotto);
+        PrizeWinningResult result = lottoMachine.checkPrizeWinning(new Lottos(firstPrizeLotto, fouthPrizeLotto), winningNumber);
 
-        assertThat(result.getRateOfReturn()).isEqualTo(1000001.0);
+        assertThat(result.getRateOfReturn()).isEqualTo(5.5);
     }
 
     @DisplayName("당첨이 되지 않으면 수익률은 0 이다")
     @Test
     void rateOfReturn3() {
         Lotto noPrizeLotto = myLotto(30, 31, 32, 33, 34, 35);
-        WinningNumber winningNumber = new WinningNumber("1, 2, 3, 4, 5, 6");
+        WinningNumber winningNumber = new WinningNumber(Arrays.asList(1, 2, 3, 4, 5, 6));
 
-        PrizeWinningResult result = lottoMachine.checkPrizeWinning(winningNumber, noPrizeLotto);
+        PrizeWinningResult result = lottoMachine.checkPrizeWinning(new Lottos(noPrizeLotto), winningNumber);
 
         assertThat(result.getRateOfReturn()).isEqualTo(0);
     }
@@ -68,16 +69,13 @@ public class LottoMachineTest {
     @DisplayName("당첨금과 로또구입금액이 같으면 수익률은 1 이다")
     @Test
     void rateOfReturn4() {
-        Lotto fourthPrizeLotto = myLotto(1, 2, 3, 10, 20, 30);
-        Lotto noPrizeLotto = myLotto(11, 12, 13, 10, 20, 30);
+        Lotto fourthPrizeLott = myLotto(1, 2, 3, 10, 20, 30);
+        WinningNumber winningNumber = new WinningNumber(Arrays.asList(1, 2, 3, 4, 5, 6));
 
-        WinningNumber winningNumber = new WinningNumber("1, 2, 3, 4, 5, 6");
-
-        PrizeWinningResult result = lottoMachine.checkPrizeWinning(winningNumber, fourthPrizeLotto, noPrizeLotto);
+        PrizeWinningResult result = lottoMachine.checkPrizeWinning(new Lottos(fourthPrizeLott), winningNumber);
 
         assertThat(result.getRateOfReturn()).isEqualTo(1);
     }
-
 
     private Lotto myLotto(int i, int i1, int i2, int i3, int i4, int i5) {
         return new Lotto(new LottoNumber(Arrays.asList(i, i1, i2, i3, i4, i5)));
